StorageContent Widget interactable
Now the widget that is spawned when a character interacts with a storage actor is fully interactable. If a slot is clicked on, the item it holds will be added to the player's inventory. Afterwards, the slot is deleted and the item removed from the storage actor's items. When all items are collected, the content widget is destroyed.