Inventory Trade Mode
Detailed Description: Now upon interacting with an NPC, the player can trade items: by pressing the "E" key when hovering over a slot in player's inventory, this item is transfered to the NPC and removed from the player's inventory. The opposite works for the case when the mouse is hovering over the NPC's inventory.