Bugfix
Detailed Description:
Fixed crashing when you press "E" when hovering over the slots with the equipped weapon/armour. Now unequip function is called and the item moved back to the inventory. Fixed stacking weapons and armour in the UI despite BaseWeapon and BaseArmour being not stackable. Functionality added: when trading, items are traded one by one instead of all of the instances per slot being transfered in one go.